requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-allauth==0.44.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-crontab==0.7.1
django-debug-toolbar==3.2.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-postgresql==0.0.3
django-rest-auth==0.9.5
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
et-xmlfile==1.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
